feat(filter): renamed to filterBy
